#535: Linux editor 消除双份 composition 状态残留清理、#[allow] 精确化、dead_code 清理、sync/config_store 移除无用 mut

+++ apps/Linux_qt/src/platform/linux_qt/utf16_converter.rs
@@ -11,7 +11,7 @@ pub fn utf16_to_utf8_offset(text: &str, utf16_offset: usize) -> usize {
         if utf16_count >= utf16_offset {
             break;
         }
-        utf16_count += ch.len_utf16() as usize;
+        utf16_count += ch.len_utf16();
         if utf16_count > utf16_offset {
             break;
         }
@@ -33,7 +33,7 @@ pub fn utf8_to_utf16_offset(text: &str, utf8_offset: usize) -> usize {
             break;
         }
         remaining -= ch_len;
-        utf16_count += ch.len_utf16() as usize;
+        utf16_count += ch.len_utf16();
     }
     utf16_count
 }
